refactor command processing
