+++ src/main/java/zos/shell/constants/Constants.java
@@ -42,7 +42,6 @@ public final class Constants {
     public static final String INVALID_PARAMETER = "invalid parameter, try again...";
     public static final String INVALID_VALUE = "invalid value, try again...";
     public static final String HIGH_QUALIFIER_ERROR = "cant change to high qualifier level, try again...";
-    public static final String NO_CONNECTIONS = "no connection(s) made or defined...";
     public static final String NO_CONNECTION_INFO = "no info, check connection settings...";
     public static final String NO_CONNECTION = "no connection to change too...";
     public static final String NO_FILES = "no files, try again...";
