+++ src/main/java/zos/shell/service/localfile/LocalFileService.java
@@ -3,8 +3,8 @@ package zos.shell.service.localfile;
 import org.apache.commons.lang3.SystemUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import zos.shell.configuration.ConfigSingleton;
 import zos.shell.constants.Constants;
+import zos.shell.singleton.configuration.ConfigSingleton;
 
 import java.io.File;
 import java.util.ArrayList;
