+++ src/main/java/zos/shell/singleton/EnvVariableSingleton.java
@@ -1,4 +1,4 @@
-package zos.shell.environment;
+package zos.shell.singleton;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
